+++ src/index.rs
@@ -32,6 +32,9 @@ mod reorg;
 mod rtx;
 mod updater;
 
+#[cfg(test)]
+pub(crate) mod testing;
+
 const SCHEMA_VERSION: u64 = 8;
 
 macro_rules! define_table {
@@ -1378,112 +1381,9 @@ impl Index {
 mod tests {
   use {
     super::*,
-    crate::runes::{Edict, Etching, RuneId, Runestone},
+    crate::index::testing::Context,
     bitcoin::secp256k1::rand::{self, RngCore},
   };
-
-  const RUNE: u128 = (21_000_000 * COIN_VALUE) as u128;
-
-  struct ContextBuilder {
-    args: Vec<OsString>,
-    chain: Chain,
-    tempdir: Option<TempDir>,
-  }
-
-  impl ContextBuilder {
-    fn build(self) -> Context {
-      self.try_build().unwrap()
-    }
-
-    fn try_build(self) -> Result<Context> {
-      let rpc_server = test_bitcoincore_rpc::builder()
-        .network(self.chain.network())
-        .build();
-
-      let tempdir = self.tempdir.unwrap_or_else(|| TempDir::new().unwrap());
-      let cookie_file = tempdir.path().join("cookie");
-      fs::write(&cookie_file, "username:password").unwrap();
-
-      let command: Vec<OsString> = vec![
-        "ord".into(),
-        "--rpc-url".into(),
-        rpc_server.url().into(),
-        "--data-dir".into(),
-        tempdir.path().into(),
-        "--cookie-file".into(),
-        cookie_file.into(),
-        format!("--chain={}", self.chain).into(),
-      ];
-
-      let options = Options::try_parse_from(command.into_iter().chain(self.args)).unwrap();
-      let index = Index::open(&options)?;
-      index.update().unwrap();
-
-      Ok(Context {
-        options,
-        rpc_server,
-        tempdir,
-        index,
-      })
-    }
-
-    fn arg(mut self, arg: impl Into<OsString>) -> Self {
-      self.args.push(arg.into());
-      self
-    }
-
-    fn args<T: Into<OsString>, I: IntoIterator<Item = T>>(mut self, args: I) -> Self {
-      self.args.extend(args.into_iter().map(|arg| arg.into()));
-      self
-    }
-
-    fn chain(mut self, chain: Chain) -> Self {
-      self.chain = chain;
-      self
-    }
-
-    fn tempdir(mut self, tempdir: TempDir) -> Self {
-      self.tempdir = Some(tempdir);
-      self
-    }
-  }
-
-  struct Context {
-    options: Options,
-    rpc_server: test_bitcoincore_rpc::Handle,
-    #[allow(unused)]
-    tempdir: TempDir,
-    index: Index,
-  }
-
-  impl Context {
-    fn builder() -> ContextBuilder {
-      ContextBuilder {
-        args: Vec::new(),
-        tempdir: None,
-        chain: Chain::Regtest,
-      }
-    }
-
-    fn mine_blocks(&self, n: u64) -> Vec<Block> {
-      let blocks = self.rpc_server.mine_blocks(n);
-      self.index.update().unwrap();
-      blocks
-    }
-
-    fn mine_blocks_with_subsidy(&self, n: u64, subsidy: u64) -> Vec<Block> {
-      let blocks = self.rpc_server.mine_blocks_with_subsidy(n, subsidy);
-      self.index.update().unwrap();
-      blocks
-    }
-
-    fn configurations() -> Vec<Context> {
-      vec![
-        Context::builder().build(),
-        Context::builder().arg("--index-sats").build(),
-      ]
-    }
-  }
 
   #[test]
   fn height_limit() {
@@ -4167,1715 +4067,5 @@ mod tests {
         .parent
         .is_none());
     }
-  }
-
-  #[test]
-  fn index_only_indexes_runes_if_flag_is_passed_and_on_mainnet() {
-    assert!(!Context::builder().build().index.has_rune_index().unwrap());
-    assert!(!Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .chain(Chain::Mainnet)
-      .build()
-      .index
-      .has_rune_index()
-      .unwrap());
-    assert!(Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build()
-      .index
-      .has_rune_index()
-      .unwrap());
-  }
-
-  #[test]
-  fn index_starts_with_no_runes() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-    assert_eq!(context.index.runes().unwrap().unwrap(), []);
-    assert_eq!(context.index.rune_balances(), []);
-  }
-
-  #[test]
-  fn empty_runestone_does_not_create_rune() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: Vec::new(),
-          etching: None,
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    assert_eq!(context.index.runes().unwrap().unwrap(), []);
-    assert_eq!(context.index.rune_balances(), []);
-  }
-
-  #[test]
-  fn etching_with_no_edicts_does_not_create_rune() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: Vec::new(),
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(0),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    assert_eq!(context.index.runes().unwrap().unwrap(), []);
-    assert_eq!(context.index.rune_balances(), []);
-  }
-
-  #[test]
-  fn etching_with_edict_creates_rune() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 0 }, vec![(id, u128::max_value())])]
-    );
-  }
-
-  #[test]
-  fn sat_corresponding_to_rune_must_have_been_mined() {
-    {
-      let context = Context::builder()
-        .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-        .build();
-
-      context.mine_blocks(1);
-
-      context.rpc_server.broadcast_tx(TransactionTemplate {
-        inputs: &[(1, 0, 0, Witness::new())],
-        op_return: Some(
-          Runestone {
-            edicts: vec![Edict {
-              id: 0,
-              amount: u128::max_value(),
-              output: 0,
-            }],
-            etching: Some(Etching {
-              decimals: 0,
-              rune: Rune(u128::from(Sat::SUPPLY - 150 * COIN_VALUE - 1)),
-            }),
-          }
-          .encipher(),
-        ),
-        ..Default::default()
-      });
-
-      context.mine_blocks(1);
-
-      assert_eq!(context.index.runes().unwrap().unwrap(), []);
-
-      assert_eq!(context.index.rune_balances(), []);
-    }
-
-    {
-      let context = Context::builder()
-        .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-        .build();
-
-      context.mine_blocks(1);
-
-      let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-        inputs: &[(1, 0, 0, Witness::new())],
-        op_return: Some(
-          Runestone {
-            edicts: vec![Edict {
-              id: 0,
-              amount: u128::max_value(),
-              output: 0,
-            }],
-            etching: Some(Etching {
-              decimals: 0,
-              rune: Rune(u128::from(Sat::SUPPLY - 150 * COIN_VALUE)),
-            }),
-          }
-          .encipher(),
-        ),
-        ..Default::default()
-      });
-
-      context.mine_blocks(1);
-
-      let id = RuneId {
-        height: 2,
-        index: 1,
-      };
-
-      assert_eq!(
-        context.index.runes().unwrap().unwrap(),
-        [(
-          id,
-          RuneEntry {
-            divisibility: 0,
-            rune: Rune(2099984997690000),
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        )]
-      );
-
-      assert_eq!(
-        context.index.rune_balances(),
-        [(OutPoint { txid, vout: 0 }, vec![(id, u128::max_value())])]
-      );
-    }
-  }
-
-  #[test]
-  fn etching_with_non_zero_decimals_and_rune() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 1,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 1,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 0 }, vec![(id, u128::max_value())])]
-    );
-  }
-
-  #[test]
-  fn allocations_over_max_supply_are_ignored() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![
-            Edict {
-              id: 0,
-              amount: u128::max_value(),
-              output: 0,
-            },
-            Edict {
-              id: 0,
-              amount: u128::max_value(),
-              output: 0,
-            },
-          ],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 0 }, vec![(id, u128::max_value())])]
-    );
-  }
-
-  #[test]
-  fn allocations_partially_over_max_supply_are_honored() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![
-            Edict {
-              id: 0,
-              amount: u128::max_value() / 2,
-              output: 0,
-            },
-            Edict {
-              id: 0,
-              amount: u128::max_value(),
-              output: 0,
-            },
-          ],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 0 }, vec![(id, u128::max_value())])]
-    );
-  }
-
-  #[test]
-  fn etching_may_allocate_less_than_max_supply() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: 100,
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: 100,
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 0 }, vec![(id, 100)])]
-    );
-  }
-
-  #[test]
-  fn etching_may_allocate_to_multiple_outputs() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![
-            Edict {
-              id: 0,
-              amount: 100,
-              output: 0,
-            },
-            Edict {
-              id: 0,
-              amount: 100,
-              output: 1,
-            },
-          ],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: 200,
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [
-        (OutPoint { txid, vout: 0 }, vec![(id, 100)]),
-        (OutPoint { txid, vout: 1 }, vec![(id, 100)])
-      ]
-    );
-  }
-
-  #[test]
-  fn allocations_to_invalid_outputs_are_ignored() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![
-            Edict {
-              id: 0,
-              amount: 100,
-              output: 0,
-            },
-            Edict {
-              id: 0,
-              amount: 100,
-              output: 3,
-            },
-          ],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: 100,
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 0 }, vec![(id, 100)]),]
-    );
-  }
-
-  #[test]
-  fn input_runes_may_be_allocated() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 0 }, vec![(id, u128::max_value())])]
-    );
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(2, 1, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: id.into(),
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: None,
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 0 }, vec![(id, u128::max_value())])]
-    );
-  }
-
-  #[test]
-  fn unallocated_runes_are_assigned_to_first_no_op_return_output() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 0 }, vec![(id, u128::max_value())])]
-    );
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(2, 1, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: Vec::new(),
-          etching: None,
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 0 }, vec![(id, u128::max_value())])]
-    );
-  }
-
-  #[test]
-  fn unallocated_runes_in_transactions_with_no_runestone_are_assigned_to_first_no_op_return_output()
-  {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 0 }, vec![(id, u128::max_value())])]
-    );
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(2, 1, 0, Witness::new())],
-      op_return: None,
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 0 }, vec![(id, u128::max_value())])]
-    );
-  }
-
-  #[test]
-  fn duplicate_runes_are_forbidden() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 0 }, vec![(id, u128::max_value())])]
-    );
-
-    context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(2, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 0 }, vec![(id, u128::max_value())])]
-    );
-  }
-
-  #[test]
-  fn outpoint_may_hold_multiple_runes() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid0 = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id0 = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id0,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(
-        OutPoint {
-          txid: txid0,
-          vout: 0
-        },
-        vec![(id0, u128::max_value())]
-      )]
-    );
-
-    let txid1 = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(2, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE + 1),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id1 = RuneId {
-      height: 3,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [
-        (
-          id0,
-          RuneEntry {
-            rune: Rune(RUNE),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        ),
-        (
-          id1,
-          RuneEntry {
-            rune: Rune(RUNE + 1),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        )
-      ]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [
-        (
-          OutPoint {
-            txid: txid1,
-            vout: 0
-          },
-          vec![(id1, u128::max_value())]
-        ),
-        (
-          OutPoint {
-            txid: txid0,
-            vout: 0
-          },
-          vec![(id0, u128::max_value())]
-        ),
-      ]
-    );
-
-    let txid2 = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(2, 1, 0, Witness::new()), (3, 1, 0, Witness::new())],
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [
-        (
-          id0,
-          RuneEntry {
-            rune: Rune(RUNE),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        ),
-        (
-          id1,
-          RuneEntry {
-            rune: Rune(RUNE + 1),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        )
-      ]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(
-        OutPoint {
-          txid: txid2,
-          vout: 0
-        },
-        vec![(id0, u128::max_value()), (id1, u128::max_value())]
-      )]
-    );
-  }
-
-  #[test]
-  fn multiple_input_runes_on_the_same_input_may_be_allocated() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid0 = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id0 = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id0,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(
-        OutPoint {
-          txid: txid0,
-          vout: 0
-        },
-        vec![(id0, u128::max_value())]
-      )]
-    );
-
-    let txid1 = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(2, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE + 1),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id1 = RuneId {
-      height: 3,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [
-        (
-          id0,
-          RuneEntry {
-            rune: Rune(RUNE),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        ),
-        (
-          id1,
-          RuneEntry {
-            rune: Rune(RUNE + 1),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        )
-      ]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [
-        (
-          OutPoint {
-            txid: txid1,
-            vout: 0
-          },
-          vec![(id1, u128::max_value())]
-        ),
-        (
-          OutPoint {
-            txid: txid0,
-            vout: 0
-          },
-          vec![(id0, u128::max_value())]
-        ),
-      ]
-    );
-
-    let txid2 = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(2, 1, 0, Witness::new()), (3, 1, 0, Witness::new())],
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [
-        (
-          id0,
-          RuneEntry {
-            rune: Rune(RUNE),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        ),
-        (
-          id1,
-          RuneEntry {
-            rune: Rune(RUNE + 1),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        )
-      ]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(
-        OutPoint {
-          txid: txid2,
-          vout: 0
-        },
-        vec![(id0, u128::max_value()), (id1, u128::max_value())]
-      )]
-    );
-
-    let txid3 = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(4, 1, 0, Witness::new())],
-      outputs: 2,
-      op_return: Some(
-        Runestone {
-          edicts: vec![
-            Edict {
-              id: id0.into(),
-              amount: u128::max_value() / 2,
-              output: 1,
-            },
-            Edict {
-              id: id1.into(),
-              amount: u128::max_value() / 2,
-              output: 1,
-            },
-          ],
-          etching: None,
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [
-        (
-          id0,
-          RuneEntry {
-            rune: Rune(RUNE),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        ),
-        (
-          id1,
-          RuneEntry {
-            rune: Rune(RUNE + 1),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        )
-      ]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [
-        (
-          OutPoint {
-            txid: txid3,
-            vout: 0
-          },
-          vec![
-            (id0, u128::max_value() / 2 + 1),
-            (id1, u128::max_value() / 2 + 1)
-          ]
-        ),
-        (
-          OutPoint {
-            txid: txid3,
-            vout: 1
-          },
-          vec![(id0, u128::max_value() / 2), (id1, u128::max_value() / 2)]
-        )
-      ]
-    );
-  }
-
-  #[test]
-  fn multiple_input_runes_on_different_inputs_may_be_allocated() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid0 = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id0 = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id0,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(
-        OutPoint {
-          txid: txid0,
-          vout: 0
-        },
-        vec![(id0, u128::max_value())]
-      )]
-    );
-
-    let txid1 = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(2, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE + 1),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id1 = RuneId {
-      height: 3,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [
-        (
-          id0,
-          RuneEntry {
-            rune: Rune(RUNE),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        ),
-        (
-          id1,
-          RuneEntry {
-            rune: Rune(RUNE + 1),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        )
-      ]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [
-        (
-          OutPoint {
-            txid: txid1,
-            vout: 0
-          },
-          vec![(id1, u128::max_value())]
-        ),
-        (
-          OutPoint {
-            txid: txid0,
-            vout: 0
-          },
-          vec![(id0, u128::max_value())]
-        )
-      ]
-    );
-
-    let txid2 = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(2, 1, 0, Witness::new()), (3, 1, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![
-            Edict {
-              id: id0.into(),
-              amount: u128::max_value(),
-              output: 0,
-            },
-            Edict {
-              id: id1.into(),
-              amount: u128::max_value(),
-              output: 0,
-            },
-          ],
-          etching: None,
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [
-        (
-          id0,
-          RuneEntry {
-            rune: Rune(RUNE),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        ),
-        (
-          id1,
-          RuneEntry {
-            rune: Rune(RUNE + 1),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        )
-      ]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(
-        OutPoint {
-          txid: txid2,
-          vout: 0
-        },
-        vec![(id0, u128::max_value()), (id1, u128::max_value())]
-      )]
-    );
-  }
-
-  #[test]
-  fn unallocated_runes_are_assigned_to_first_no_op_return_output_when_op_return_is_not_last_output()
-  {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 0 }, vec![(id, u128::max_value())])]
-    );
-
-    let txid = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(2, 1, 0, Witness::new())],
-      op_return: Some(
-        script::Builder::new()
-          .push_opcode(opcodes::all::OP_RETURN)
-          .into_script(),
-      ),
-      op_return_index: Some(0),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [(
-        id,
-        RuneEntry {
-          rune: Rune(RUNE),
-          divisibility: 0,
-          supply: u128::max_value(),
-          rarity: Rarity::Uncommon,
-        }
-      )]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [(OutPoint { txid, vout: 1 }, vec![(id, u128::max_value())])]
-    );
-  }
-
-  #[test]
-  fn rune_rarity_is_assigned_correctly() {
-    let context = Context::builder()
-      .arg("--index-runes-pre-alpha-i-agree-to-get-rekt")
-      .build();
-
-    context.mine_blocks(1);
-
-    let txid0 = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    let id0 = RuneId {
-      height: 2,
-      index: 1,
-    };
-
-    let txid1 = context.rpc_server.broadcast_tx(TransactionTemplate {
-      inputs: &[(1, 0, 0, Witness::new())],
-      op_return: Some(
-        Runestone {
-          edicts: vec![Edict {
-            id: 0,
-            amount: u128::max_value(),
-            output: 0,
-          }],
-          etching: Some(Etching {
-            decimals: 0,
-            rune: Rune(RUNE + 1),
-          }),
-        }
-        .encipher(),
-      ),
-      ..Default::default()
-    });
-
-    context.mine_blocks(1);
-
-    let id1 = RuneId {
-      height: 2,
-      index: 2,
-    };
-
-    assert_eq!(
-      context.index.runes().unwrap().unwrap(),
-      [
-        (
-          id0,
-          RuneEntry {
-            rune: Rune(RUNE),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Uncommon,
-          }
-        ),
-        (
-          id1,
-          RuneEntry {
-            rune: Rune(RUNE + 1),
-            divisibility: 0,
-            supply: u128::max_value(),
-            rarity: Rarity::Common,
-          }
-        ),
-      ]
-    );
-
-    assert_eq!(
-      context.index.rune_balances(),
-      [
-        (
-          OutPoint {
-            txid: txid0,
-            vout: 0
-          },
-          vec![(id0, u128::max_value())]
-        ),
-        (
-          OutPoint {
-            txid: txid1,
-            vout: 0
-          },
-          vec![(id1, u128::max_value())]
-        ),
-      ]
-    );
   }
 }
